split case runners into separate files
